feat(telegram): receive and surface user message reactions (#10075)
